update v 1.1.3

+++ bmmodel/apply/apply.go
@@ -15,7 +15,7 @@ type BmApply struct {
 	Status     float64           `json:"status" bson:"status"` //0=未处理，1=已处理
 	ApplyTime  float64           `json:"apply_time" bson:"apply_time"`
 	ExceptTime float64           `json:"except_time" bson:"except_time"`
-	CreateTime int64             `json:"create_time" bson:"create_time"`
+	CreateTime float64             `json:"create_time" bson:"create_time"`
 	BrandId    string            `json:"brandId" bson:"brandId"`
 	ApplyFrom  string            `json:"applyFrom" bson:"applyFrom"`
 	CourseType float64           `json:"courseType" bson:"courseType"` //0活动 1体验课 2普通课程 -1预注册
