+++ bmmodel/reservable/reservable.go
@@ -16,7 +16,7 @@ type BmReservable struct {
 	Status     float64 `json:"status" bson:"status"` //0活动 1体验课 2普通课程
 	StartDate  float64 `json:"start_date" bson:"start_date"`
 	EndDate    float64 `json:"end_date" bson:"end_date"`
-	CreateTime int64   `json:"create_time" bson:"create_time"`
+	CreateTime float64   `json:"create_time" bson:"create_time"`
 
 	SessionInfo sessioninfo.BmSessionInfo `json:"SessionInfo" jsonapi:"relationships"`
 }
